refactor: inline the animation check into notification
Replacing it with getStateAnimations() does not work. The notification card declares
the same keyframes for the opening and the closing state, so once the opening
animation has ended nothing is restarted at the moment the closing state is applied,
and the animations are not observable when the decision has to be made. Measured in
WebKit and Firefox, where the card takes the max-height fallback path: the computed
style reports both keyframe names and a 0.1s duration, while getAnimations() returns
only a transition.

Inline the predicate as a private method instead, which lets the shared
shouldAnimate() helper go away along with its tests. Reading the animation objects
here needs the card to use distinct keyframes per state first.
