+++ packages/overlay/src/vaadin-overlay-utils.js
@@ -86,24 +86,6 @@ export function observeMove(element, callback) {
 }
 
 /**
- * Detect whether an animation runs on the given element, so that its end can be
- * awaited before the element is hidden or removed. An element that is not rendered,
- * has no animation name, or has a zero duration does not fire `animationend`.
- *
- * @param {HTMLElement} element
- * @return {boolean}
- */
-export function shouldAnimate(element) {
-  const style = getComputedStyle(element);
-  const name = style.getPropertyValue('animation-name');
-  const hasDuration = style
-    .getPropertyValue('animation-duration')
-    .split(',')
-    .some((duration) => parseFloat(duration) > 0);
-  return style.getPropertyValue('display') !== 'none' && name && name !== 'none' && hasDuration;
-}
-
-/**
  * Collect the animations that report the state of the given element, so that their end can be
  * awaited before the element is hidden or removed. The animations are read instead of the
  * computed style, so that the decision to wait cannot disagree with what the browser actually
