+++ packages/overlay/test/utils.test.js
@@ -1,6 +1,6 @@
 import { expect } from '@vaadin/chai-plugins';
 import { fixtureSync } from '@vaadin/testing-helpers';
-import { getStateAnimations, shouldAnimate } from '../src/vaadin-overlay-utils.js';
+import { getStateAnimations } from '../src/vaadin-overlay-utils.js';
 
 const styles = document.createElement('style');
 styles.textContent = `
@@ -61,35 +61,5 @@ describe('getStateAnimations', () => {
     child.style.animation = 'utils-fade 1s';
     element.appendChild(child);
     expect(getStateAnimations(element)).to.be.empty;
-  });
-});
-
-describe('shouldAnimate', () => {
-  let element;
-
-  beforeEach(() => {
-    element = fixtureSync('<div></div>');
-  });
-
-  it('should return false when the element is not rendered', () => {
-    element.style.animation = 'utils-fade 1s';
-    element.style.display = 'none';
-    expect(shouldAnimate(element)).to.be.false;
-  });
-
-  it('should return false when the element has no animation name', () => {
-    element.style.animationDuration = '1s';
-    expect(shouldAnimate(element)).to.be.false;
-  });
-
-  it('should return false when the animation duration is zero', () => {
-    element.style.animation = 'utils-fade 0s';
-    expect(shouldAnimate(element)).to.be.false;
-  });
-
-  it('should return true when one of the animation durations is not zero', () => {
-    element.style.animationName = 'utils-fade, utils-scale';
-    element.style.animationDuration = '0s, 1s';
-    expect(shouldAnimate(element)).to.be.true;
   });
 });
